app/requirements.txt
spacy
weaviate-client==4.4.4
sentence-transformers